pkg/cdi: add K8s annotation handling functions.
Add functions for annotating containers for CDI device injection
and for extracting CDI device names from such annotations. These
functions can be used by CDI-aware devices plugins and runtimes,
to request the injection of CDI devices and to recover the names
of devices to inject correspondingly.

For example, the following call to AnnotateInjection
    annotations := map[string]string{}
    AnnotateInjection(annotations, "vendor.gpu", []string{
        "vendor.com/gpu=gpu0",
        "vendor.com/gpu=gpu1",
    })
produces the following annotations
    map[string]string{
        "cdi.k8s.io/vendor.gpu": "vendor.com/gpu=gpu0,vendor.com/gpu=gpu1",
    }

Subsequently parsing these annotations with
    devices, err := ParseAnnotations(annotations)
extracts and returns the original slice of devices to inject:
    []string{
        "vendor.com/gpu=gpu0",
        "vendor.com/gpu=gpu1",
    }

Notes:
  These functions provide a temporary solution for delivering
  CDI device injection requests from either a client or a K8s
  Device Plugin to the container runtime. They can be removed
  once dedicated fields to the same effect are added natively
  to the CRI protocol.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>