pkg/cdi: add ResolveDevices for runtimes.
Add ResolveDevices to handle devices injection in runtimes.
Unlike InjectDevices which is intended for injection on the
client side, ResolveDevices gives more control over details
of the injected data. The runtime can provide an OciHandler
with functions for modifying and tracking OCI data prior to
injection. The runtime can use this to adjust data and post-
process an updated OCI Spec as necessary.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>